Made modifications to server side so client side won't use as many api requests
