+++ src/.env
@@ -1,0 +1,1 @@
+APIKEY='10b2e5a21cmshbad5454826f4cb8p112f65jsn6fe0bd9f4951'
